feat(modal/formatter): replace source image in text to chrome.runtime local path
